Adds form example to sidebar in storybook and removes Demo page
