--- components/src/DemoPage/index.js
@@ -1,1 +1,0 @@
-export { default as DemoPage } from "./DemoPage";
