+++ components/src/pages/Details.story.js
@@ -18,7 +18,15 @@ import {
   Tab,
   Table,
   Form,
+  FormSection,
+  Input,
+  Select,
   Textarea,
+  CheckboxGroup,
+  Checkbox,
+  RadioGroup,
+  Radio,
+  Toggle,
   PrimaryButton,
   QuietButton
 } from "../index";
@@ -124,6 +132,8 @@ const productionRows = [
     id: "r8"
   }
 ];
+
+const options = [{ value: "planned", label: "Planned" }, { value: "booked", label: "Booked" }];
 
 const StorybookContentHeight = "calc(100vh - 72px - 48px - 16px)";
 const StorybookSidebarWidth = "calc(100vw - 48px - 16px)";
@@ -294,12 +304,69 @@ storiesOf("Pages/Details page", module)
             borderLeft={{ medium: `solid 1px ${theme.colors.lightGrey}` }}
             position={{ extraSmall: "absolute", medium: "static" }}
             borderRadius="medium"
+            //height={StorybookContentHeight}
+            //overflow="auto"
           >
             <Flex justifyContent="space-between" alignItems="flex-start" mb="x4">
-              <SectionTitle mt="half">I am sidebar</SectionTitle>
+              <SectionTitle mt="half">Job 324400</SectionTitle>
               <IconicButton icon="close" />
             </Flex>
-            <Text>I am sidbar content.</Text>
+            <Form>
+              <FormSection title="Job Information">
+                <Input id="project" labelText="Project" placeholder="Project 128703" />
+                <Input
+                  id="project-description"
+                  labelText="Project description"
+                  requirementText="(Optional)"
+                  helpText="Project description helps identify the project."
+                />
+                <Select id="project-status" labelText="Project status" options={options} />
+                <Input
+                  id="item-code"
+                  labelText="Item code"
+                  defaultValue="WS2SB6"
+                  errorMessage="Item WS2SB6 does not exist."
+                />
+                <Input id="eaches-expected" labelText="Eaches expected on Job" placeholder="2 000" />
+                <Input id="eaches-remaining" labelText="Eaches remaining on Project" defaultValue="18 000" disabled />
+                <Input id="scheduled-start" labelText="Scheduled start" placeholder="MMM-DD-YYYY" />
+                <Input id="scheduled-end" labelText="Scheduled end" placeholder="MMM-DD-YYYY" />
+                <CheckboxGroup labelText="Line Lead" name="linelead" requirementText="(Optional)">
+                  <Checkbox value="christiaan" labelText="Christiaan Oostenbrug" />
+                  <Checkbox value="matt" labelText="Matt Dunn" />
+                  <Checkbox value="clemens" labelText="Clemens Park" disabled checked />
+                  <Checkbox value="nikola" labelText="Nikola Pejcic" disabled />
+                </CheckboxGroup>
+                <RadioGroup
+                  errorMessage="Only yes can be selected..."
+                  labelText="Reconcile"
+                  name="settingSelection"
+                  defaultValue="yes"
+                  id="reconcile"
+                >
+                  <Radio value="yes" labelText="Yes" />
+                  <Radio value="no" labelText="No" />
+                  <Radio value="maybe" labelText="Maybe" disabled />
+                </RadioGroup>
+                <Toggle id="job-visibility" labelText="Job Visibility" onText="Visible" offText="Hidden" />
+              </FormSection>
+              <FormSection title="Rejects">
+                <Input
+                  defaultValue="235432"
+                  id="items"
+                  labelText="Item"
+                  errorMessage="Item 235432 is not a valid entry."
+                />
+                <Input id="quantity" labelText="Quantity" />
+                <Toggle
+                  id="reject-visibility"
+                  labelText="Reject visibility"
+                  onText="Visible"
+                  offText="Hidden"
+                  disabled
+                />
+              </FormSection>
+            </Form>
           </Box>
         </Flex>
       </Flex>
@@ -407,7 +474,7 @@ storiesOf("Pages/Details page", module)
               <SectionTitle mt="half">I am sidebar</SectionTitle>
               <IconicButton icon="close" />
             </Flex>
-            <Text>I am sidbar content.</Text>
+            <Text>I am sidebar content.</Text>
           </Box>
         </Flex>
       </Flex>
